Win64 port

+++ src/GClasses/GActivation.h
@@ -46,14 +46,14 @@ public:
 	virtual const char* name() const = 0;
 
 	/// The activation function
-	virtual double squash(double x, int index) = 0;
+	virtual double squash(double x, size_t index) = 0;
 
 	/// The derivative of the activation function
-	virtual double derivative(double x, int index) = 0;
+	virtual double derivative(double x, size_t index) = 0;
 
 	/// The inverse of the activation function. (This function may throw an exception
 	/// if the activation function cannot be inverted.)
-	virtual double inverse(double y, int index) = 0;
+	virtual double inverse(double y, size_t index) = 0;
 
 	/// Returns a clone of this object
 	virtual GActivationFunction* clone() = 0;
@@ -61,7 +61,7 @@ public:
 	/// This computes the derivative of the net value. (Sometimes, such as with
 	/// GActivationLogistic, it is more efficient to compute this from the activation
 	/// value, so both are provided.)
-	virtual double derivativeOfNet(double net, double activation, int index) { return derivative(net, index); }
+	virtual double derivativeOfNet(double net, double activation, size_t index) { return derivative(net, index); }
 
 	/// Serialize this object
 	virtual GDomNode* serialize(GDom* pDoc) const;
@@ -90,7 +90,7 @@ public:
 	virtual const char* name() const { return "logistic"; }
 
 	/// The logistic function. Returns 1.0/(e^(-x)+1.0)
-	virtual double squash(double x, int index)
+	virtual double squash(double x, size_t index)
 	{
 		if(x >= 700.0) // Don't trigger a floating point exception
 			return 1.0;
@@ -100,10 +100,10 @@ public:
 	}
 
 	/// Returns d*(1.0-d), where d=squash(x)
-	virtual double derivative(double x, int index) { double d = squash(x, index); return d * (1.0 - d); }
+	virtual double derivative(double x, size_t index) { double d = squash(x, index); return d * (1.0 - d); }
 
 	/// The logit function. Returns log(y)-log(1.0-y)
-	virtual double inverse(double y, int index)
+	virtual double inverse(double y, size_t index)
 	{
 		if(y >= 1.0)
 			return 700.0;
@@ -113,7 +113,7 @@ public:
 	}
 
 	/// Returns y*(1.0-y)
-	virtual double derivativeOfNet(double net, double activation, int index) { return activation * (1.0 - activation); }
+	virtual double derivativeOfNet(double net, double activation, size_t index) { return activation * (1.0 - activation); }
 
 	/// See the comment for GActivationFunction::clone
 	virtual GActivationFunction* clone() { return new GActivationLogistic(); }
@@ -127,13 +127,13 @@ public:
 	virtual const char* name() const { return "arctan"; }
 
 	/// Returns atan(x). The result will be in the range -PI/2 <= y <= PI/2
-	virtual double squash(double x, int index) { return atan(x); }
+	virtual double squash(double x, size_t index) { return atan(x); }
 
 	/// Returns 1/(x*x+1.0)
-	virtual double derivative(double x, int index) { return 1.0 / (x * x + 1.0); }
+	virtual double derivative(double x, size_t index) { return 1.0 / (x * x + 1.0); }
 
 	/// Returns tan(y), where -PI/2 <= y <= PI/2
-	virtual double inverse(double y, int index) { return tan(y); }
+	virtual double inverse(double y, size_t index) { return tan(y); }
 
 	/// See the comment for GActivationFunction::clone
 	virtual GActivationFunction* clone() { return new GActivationArcTan(); }
@@ -147,7 +147,7 @@ public:
 	virtual const char* name() const { return "tanh"; }
 
 	/// Returns tanh(x). The result is in the range -1 <= y <= 1
-	virtual double squash(double x, int index)
+	virtual double squash(double x, size_t index)
 	{
 		//return tanh(x);
 		if(x >= 700.0)
@@ -160,7 +160,7 @@ public:
 	}
 
 	/// Returns sech(x)*sech(x)
-	virtual double derivative(double x, int index)
+	virtual double derivative(double x, size_t index)
 	{
 		if(x >= 700.0)
 			return 1.0;
@@ -173,7 +173,7 @@ public:
 	}
 
 	/// Returns atanh(y), where -1 <= y <= 1
-	virtual double inverse(double y, int index)
+	virtual double inverse(double y, size_t index)
 	{
 #ifdef WINDOWS
 		return 0.5 * (log(1.0 + y) - log(1.0 - y));
@@ -197,13 +197,13 @@ public:
 	virtual const char* name() const { return "algebraic"; }
 
 	/// Returns x/(sqrt(x*x+1.0). The result is in the range -1 <= y <= 1
-	virtual double squash(double x, int index) { return x / (sqrt(x * x + 1.0)); }
+	virtual double squash(double x, size_t index) { return x / (sqrt(x * x + 1.0)); }
 
 	/// Returns 1.0/(sqrt(x*x+1))-(x*x)/pow(x*x+1,1.5)
-	virtual double derivative(double x, int index) { return 1.0 / (sqrt(x * x + 1)) - (x * x) / pow(x * x + 1, 1.5); }
+	virtual double derivative(double x, size_t index) { return 1.0 / (sqrt(x * x + 1)) - (x * x) / pow(x * x + 1, 1.5); }
 
 	/// Returns y / (sqrt(1.0 - (y * y)))
-	virtual double inverse(double y, int index) { return y / (sqrt(1.0 - (y * y))); }
+	virtual double inverse(double y, size_t index) { return y / (sqrt(1.0 - (y * y))); }
 
 	/// See the comment for GActivationFunction::clone
 	virtual GActivationFunction* clone() { return new GActivationAlgebraic(); }
@@ -220,16 +220,16 @@ public:
 	virtual const char* name() const { return "identity"; }
 
 	/// Returns x
-	virtual double squash(double x, int index) { return x; }
+	virtual double squash(double x, size_t index) { return x; }
 
 	/// Returns 1.0
-	virtual double derivative(double x, int index) { return 1.0; }
+	virtual double derivative(double x, size_t index) { return 1.0; }
 
 	/// Returns y
-	virtual double inverse(double y, int index) { return y; }
+	virtual double inverse(double y, size_t index) { return y; }
 
 	/// Returns 1.0
-	virtual double derivativeOfNet(double net, double activation, int index) { return 1.0; }
+	virtual double derivativeOfNet(double net, double activation, size_t index) { return 1.0; }
 
 	/// See the comment for GActivationFunction::clone
 	virtual GActivationFunction* clone() { return new GActivationIdentity(); }
@@ -250,19 +250,19 @@ public:
 	virtual const char* name() const { return "bend"; }
 
 	/// Returns the bend function of x
-	virtual double squash(double x, int index)
+	virtual double squash(double x, size_t index)
 	{
 		return BEND_AMOUNT * (sqrt(x * x + 1) - 1) + x;
 	}
 
 	/// Returns the derivative of the bend function
-	virtual double derivative(double x, int index)
+	virtual double derivative(double x, size_t index)
 	{
 		return BEND_AMOUNT * x / sqrt(x * x + 1) + 1;
 	}
 
 	/// Returns the inverse of the bend function
-	virtual double inverse(double y, int index)
+	virtual double inverse(double y, size_t index)
 	{
 		return (BEND_AMOUNT * sqrt(y * y + 2.0 * BEND_AMOUNT * y + 1) - y - BEND_AMOUNT) / (BEND_AMOUNT * BEND_AMOUNT - 1);
 	}
@@ -296,19 +296,19 @@ public:
 	virtual GDomNode* serialize(GDom* pDoc) const;
 
 	/// Returns the bend function of x
-	virtual double squash(double x, int index)
+	virtual double squash(double x, size_t index)
 	{
 		return m_hinges.v[index] * (sqrt(x * x + 1) - 1) + x;
 	}
 
 	/// Returns the derivative of the bend function
-	virtual double derivative(double x, int index)
+	virtual double derivative(double x, size_t index)
 	{
 		return m_hinges.v[index] * x / sqrt(x * x + 1) + 1;
 	}
 
 	/// Returns the inverse of the bend function
-	virtual double inverse(double y, int index)
+	virtual double inverse(double y, size_t index)
 	{
 		double v = m_hinges.v[index];
 		return (v * sqrt(y * y + 2.0 * v * y + 1) - y - v) / (v * v - 1);
@@ -353,12 +353,12 @@ public:
 	virtual GDomNode* serialize(GDom* pDoc) const;
 
 	/// Returns the logexp function of x with the parameterized alpha value
-	virtual double squash(double x, int index)
+	virtual double squash(double x, size_t index)
 	{
 		return GMath::logExp(m_alphas.v[index], x);
 	}
 
-	virtual double derivativeOfNet(double net, double activation, int index)
+	virtual double derivativeOfNet(double net, double activation, size_t index)
 	{
 		return 1.0;
 /*
@@ -372,7 +372,7 @@ public:
 	}
 
 	/// Returns the derivative of the bend function
-	virtual double derivative(double x, int index)
+	virtual double derivative(double x, size_t index)
 	{
 		double a = m_alphas.v[index];
 		if(a >= 0)
@@ -385,7 +385,7 @@ public:
 	}
 
 	/// Returns the inverse of the bend function
-	virtual double inverse(double y, int index)
+	virtual double inverse(double y, size_t index)
 	{
 		double a = m_alphas.v[index];
 		return GMath::logExp(-a, y);
@@ -418,13 +418,13 @@ public:
 	/// Returns the name of this activation function
 	virtual const char* name() const { return "bidir"; }
 
-	virtual double squash(double x, int index)
+	virtual double squash(double x, size_t index)
 	{
 		double d = sqrt(x * x + 1.0);
 		return sqrt(d + x) - sqrt(d - x);
 	}
 
-	virtual double derivative(double x, int index)
+	virtual double derivative(double x, size_t index)
 	{
 		if(std::abs(x) > 1e7)
 			return 0.0;
@@ -433,7 +433,7 @@ public:
 		return (t + 1.0) / (2.0 * sqrt(d + x)) - (t - 1.0) / (2.0 * sqrt(d - x));
 	}
 
-	virtual double inverse(double y, int index)
+	virtual double inverse(double y, size_t index)
 	{
 		double d = y * y;
 		if(y >= 0.0)
@@ -453,11 +453,11 @@ public:
 	/// Returns the name of this activation function
 	virtual const char* name() const { return "gaussian"; }
 
-	virtual double squash(double x, int index) { return exp(-(x * x)); }
+	virtual double squash(double x, size_t index) { return exp(-(x * x)); }
 
-	virtual double derivative(double x, int index) { return -2.0 * x * exp(-(x * x)); }
+	virtual double derivative(double x, size_t index) { return -2.0 * x * exp(-(x * x)); }
 
-	virtual double inverse(double y, int index)
+	virtual double inverse(double y, size_t index)
 	{
 		throw Ex("Not invertible");
 		return 0;
@@ -476,11 +476,11 @@ public:
 	/// Returns the name of this activation function
 	virtual const char* name() const { return "sin"; }
 
-	virtual double squash(double x, int index) { return sin(x); }
+	virtual double squash(double x, size_t index) { return sin(x); }
 
-	virtual double derivative(double x, int index) { return cos(x); }
+	virtual double derivative(double x, size_t index) { return cos(x); }
 
-	virtual double inverse(double y, int index) { return asin(y); }
+	virtual double inverse(double y, size_t index) { return asin(y); }
 
 	/// See the comment for GActivationFunction::clone
 	virtual GActivationFunction* clone() { return new GActivationSin(); }
@@ -495,11 +495,11 @@ public:
 	/// Returns the name of this activation function
 	virtual const char* name() const { return "sinc"; }
 
-	virtual double squash(double x, int index) { return x == 0 ? 1.0 : sin(x) / x; }
+	virtual double squash(double x, size_t index) { return x == 0 ? 1.0 : sin(x) / x; }
 
-	virtual double derivative(double x, int index) { return x == 0 ? 0.0 : cos(x) / x - sin(x) / (x * x); }
+	virtual double derivative(double x, size_t index) { return x == 0 ? 0.0 : cos(x) / x - sin(x) / (x * x); }
 
-	virtual double inverse(double y, int index)
+	virtual double inverse(double y, size_t index)
 	{
 		throw Ex("The Sinc function cannot be inverted");
 		return 0;
@@ -518,7 +518,7 @@ public:
 	virtual const char* name() const { return "logisticderiv"; }
 
 	/// The derivative of the logistic function.
-	virtual double squash(double x, int index)
+	virtual double squash(double x, size_t index)
 	{
 		if(x >= 700.0) // Don't trigger a floating point exception
 			return 0.0;
@@ -529,16 +529,16 @@ public:
 	}
 
 	/// Returns d*(1.0-d), where d=squash(x)
-	virtual double derivative(double x, int index) { double d = squash(x, index); return d * (1.0 - d) * (1.0 - 2.0 * d); }
+	virtual double derivative(double x, size_t index) { double d = squash(x, index); return d * (1.0 - d) * (1.0 - 2.0 * d); }
 
 	/// The logit function. Returns log(y)-log(1.0-y)
-	virtual double inverse(double y, int index)
+	virtual double inverse(double y, size_t index)
 	{
 		throw Ex("This function is not easily invertible");
 	}
 
 	/// Returns y*(1.0-y)
-	virtual double derivativeOfNet(double net, double activation, int index)
+	virtual double derivativeOfNet(double net, double activation, size_t index)
 	{
 		double t = 1.0 - 2.0 / (exp(-net) + 1.0);
 		return activation * t;
@@ -556,11 +556,11 @@ public:
 	/// Returns the name of this activation function
 	virtual const char* name() const { return "relu"; }
 
-	virtual double squash(double x, int index) { return std::max(0.0, x); }
+	virtual double squash(double x, size_t index) { return std::max(0.0, x); }
 
-	virtual double derivative(double x, int index) { return (x >= 0.0 ? 1.0 : 0.0); }
+	virtual double derivative(double x, size_t index) { return (x >= 0.0 ? 1.0 : 0.0); }
 
-	virtual double inverse(double y, int index) { return y; }
+	virtual double inverse(double y, size_t index) { return y; }
 
 	/// See the comment for GActivationFunction::clone
 	virtual GActivationFunction* clone() { return new GActivationRectifiedLinear(); }
@@ -574,11 +574,11 @@ public:
 	/// Returns the name of this activation function
 	virtual const char* name() const { return "softplus"; }
 
-	virtual double squash(double x, int index) { return x > 500 ? x : log(1.0 + exp(x)); }
+	virtual double squash(double x, size_t index) { return x > 500 ? x : log(1.0 + exp(x)); }
 
-	virtual double derivative(double x, int index) { return 1.0 / (1.0 + exp(-x)); }
+	virtual double derivative(double x, size_t index) { return 1.0 / (1.0 + exp(-x)); }
 
-	virtual double inverse(double y, int index) { return log(exp(y) - 1.0); }
+	virtual double inverse(double y, size_t index) { return log(exp(y) - 1.0); }
 
 	/// See the comment for GActivationFunction::clone
 	virtual GActivationFunction* clone() { return new GActivationSoftPlus(); }
@@ -592,11 +592,11 @@ public:
 	/// Returns the name of this activation function
 	virtual const char* name() const { return "softplus2"; }
 
-	virtual double squash(double x, int index) { return 0.5 * (sqrt(x * x + 1) + x); }
+	virtual double squash(double x, size_t index) { return 0.5 * (sqrt(x * x + 1) + x); }
 
-	virtual double derivative(double x, int index) { return 0.5 * (x / sqrt(x * x + 1) + 1.0); }
+	virtual double derivative(double x, size_t index) { return 0.5 * (x / sqrt(x * x + 1) + 1.0); }
 
-	virtual double inverse(double y, int index) { return (2 * y - 1) * (2 * y + 1) / (4 * y); }
+	virtual double inverse(double y, size_t index) { return (2 * y - 1) * (2 * y + 1) / (4 * y); }
 
 	/// See the comment for GActivationFunction::clone
 	virtual GActivationFunction* clone() { return new GActivationSoftPlus2(); }
